% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fload_packages.R
\name{f_load_packages}
\alias{f_load_packages}
\title{Install and Load Multiple R Packages}
\usage{
f_load_packages(...)
}
\arguments{
\item{...}{Unquoted or quoted names of packages to be installed and loaded. These should be valid package names available on CRAN.}
}
\value{
None. The function is called for its side effects of installing and loading packages.
}
\description{
Checks if the specified packages are installed. If not, it installs them and then loads them into the global R session.
}
\details{
The function takes a list or vector indicating package names, installs any that are missing, and loads all specified packages into the global environment of the R session. It uses \code{requireNamespace()} to check for installation and \code{library()} to load the packages.
}
\author{
Sander H. van Delden  \email{plantmind@proton.me} \cr
}
