% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flong.R
\name{plot.f_long}
\alias{plot.f_long}
\title{Plot method for f_long objects}
\usage{
\method{plot}{f_long}(x, summary = TRUE, ...)
}
\arguments{
\item{x}{An object of class \code{f_long} (output from \code{f_long}).}

\item{summary}{Logical. If \code{TRUE}, generates the summary table within the scan. Default is \code{TRUE}.}

\item{...}{Additional arguments passed to \code{f_scan}.}
}
\value{
Returns the output of \code{f_scan} (an object of class \code{f_scan}) invisibly.
}
\description{
Automatically runs a \code{f_scan} diagnostic plot on data created by \code{f_long}.
}
