% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstat_wizard.R
\name{print.f_stat_wizard}
\alias{print.f_stat_wizard}
\title{Print method for f_stat_wizard}
\usage{
\method{print}{f_stat_wizard}(x, ...)
}
\arguments{
\item{x}{An object of class \code{f_stat_wizard}.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print method for f_stat_wizard
}
