% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flong.R
\name{summary.f_long}
\alias{summary.f_long}
\title{Summary method for f_long objects}
\usage{
\method{summary}{f_long}(object, ...)
}
\arguments{
\item{object}{An object of class \code{f_long} (output from \code{f_long}).}

\item{...}{Additional arguments passed to \code{f_summary}.}
}
\value{
Returns the summary table (usually a data frame or tibble) produced by \code{f_summary}.
}
\description{
Automatically runs the \code{f_summary} function on data created by \code{f_long}
using the attributes stored in the object.
}
