% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foutliers.R
\name{print.f_outliers}
\alias{print.f_outliers}
\title{Print method for f_outliers objects}
\usage{
\method{print}{f_outliers}(
  x,
  col_width = 6,
  table_width = 90,
  digits = 2,
  allow_integer_decimal_mix = FALSE,
  ...
)
}
\arguments{
\item{x}{Object of class f_outliers.}

\item{col_width}{Integer. Max characters in header before line break. Default \code{6}.}

\item{table_width}{Integer or \code{NULL}. Characters after which table splits. Default \code{90}.}

\item{digits}{Integer. Number of decimal digits to use in formatting. Default is \code{3}.}

\item{allow_integer_decimal_mix}{Logical. If \code{TRUE}, each individual cell is evaluated: integer values are displayed without decimal places, and non-integer values are displayed with the specified number of decimal places, i.e. \code{digits}. Default is \code{FALSE}, when a column contains a mix of integers and decimal values, all values are displayed with the specified number of decimal places. Note: columns containing only integers are **always** displayed without decimal places, regardless of \code{allow_integer_decimal_mix}.}

\item{...}{Additional arguments passed to \code{pander}.}
}
\value{
Invisibly returns \code{1}.
}
\description{
Prints a formatted summary table to the console.
}
