% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstat_wizard.R
\name{print.f_stat_wizard}
\alias{print.f_stat_wizard}
\title{Print method for f_stat_wizard}
\usage{
\method{print}{f_stat_wizard}(x, plots = TRUE, ...)
}
\arguments{
\item{x}{An object of class \code{f_stat_wizard}.}

\item{plots}{Logical. If \code{TRUE}, display diagnostic plots (histogram and
QQ-plot) if they were generated with \code{plots=TRUE}. Default \code{TRUE}.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print method for f_stat_wizard
}
