% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsummary.R
\name{print.f_summary}
\alias{print.f_summary}
\title{Print method for f_summary objects}
\usage{
\method{print}{f_summary}(x, col_width = 6, table_width = 90, ...)
}
\arguments{
\item{x}{Object of class f_summary}

\item{col_width}{Integer. Specifies the maximum number of characters allowed in table header columns before a line break is inserted. Defaults to \code{6}.}

\item{table_width}{Integer or \code{NULL}. Defines the number of characters after which the table is split into separate sections. Defaults to \code{NULL}, meaning no break is applied.}

\item{...}{Additional arguments passed to the \code{pander} function.}
}
\value{
This function is called for its side effect of printing a formatted output to the console
and does not return a useful value. It invisibly returns \code{1}.
}
\description{
This function prints \code{f_summary} objects.
}
